match-startページのコンポーネントが、マージの課程で２つできていたから修正
